requirements.txt
langchain==0.3.0
langchain-core==0.3.1
langchain-openai==0.2.0
langchain-text-splitters==0.3.0
pandas==2.2.2
python-dotenv==1.0.1
graphviz==0.20.3
scikit-learn==1.6.1
openai==1.47.1
httpx==0.27.2
bottleneck==1.3.6
sentence_transformers==3.4.1
numexpr==2.8.4
treelib==1.7.1